upgraded entt version, going to replacae my serial stuff with entt::meta because it's much more featured

+++ src/impl/engine/io.cpp
@@ -1,4 +1,4 @@
-#include "io.h"
+#include "engine/io.h"
 #include "engine/script.h"
 
 #include <filesystem>
@@ -59,19 +59,23 @@ std::string CopyToHotDll(const std::string& filepath)
 #ifdef WIN32
 
     // append a unique id to the pdb because devenv never releases the file handle
-    std::stringstream ss; ss << "." << pdbId << ".pdb";
     
     path src_pdb = src.parent_path() / src.filename().replace_extension(".pdb");
-    path dst_pdb = "./hot/" / src_pdb.filename().replace_extension(ss.str());
 
-    // simple copy for pdb
-    copy_file(src_pdb, dst_pdb, copy_options::overwrite_existing);
+    if (exists(src_pdb))
+    {
+        std::stringstream ss; ss << "." << pdbId << ".pdb";
+        path dst_pdb = "./hot/" / src_pdb.filename().replace_extension(ss.str());
+    
+        // simple copy for pdb
+        copy_file(src_pdb, dst_pdb, copy_options::overwrite_existing);
 
-    // replace pdb link in dll
-    std::string _;
-    cr_pdb_replace(dst.string(), dst_pdb.string(), _);
+        // replace pdb link in dll
+        std::string _;
+        cr_pdb_replace(dst.string(), dst_pdb.string(), _);
 
-    pdbId += 1;
+        pdbId += 1;
+    }
 #endif
 
     return dst.string();
@@ -136,16 +140,21 @@ funcptr<_r, _a...> wLoadFunction(void* lib, const std::string& path)
     return load;
 }
 
-void PassContextsToDll(void* lib, Application& app)
+bool PassContextsToDll(void* lib, Application& app)
 {
     ScriptingContext ctx;
     ctx.ctx_imgui  = app.GetWindow().GetImGuiContext();
     ctx.ctx_opengl = app.GetWindow().GetGLContext();
     ctx.ctx_window = app.GetWindow().GetSDLWindow();
     ctx.ctx_time   = Time::GetContext();
-    
+    ctx.ctx_render = Render::GetContext();
+
     auto fn = wLoadFunction<void, ScriptingContext>(lib, "PassContexts");
-    fn(ctx);
+    
+    if (fn) fn(ctx);
+    else log_io("e~Failed to pass context to library");
+
+    return !!fn;
 }
 
 void CallOnAllSystems(void* lib, World* world, const WorldFileData& data, const std::string& what)
@@ -158,7 +167,9 @@ void CallOnAllSystems(void* lib, World* world, const WorldFileData& data, const 
         // soft fail
 
         auto load = wLoadFunction<void, World*>(lib, ss.str());
+        
         if (load) load(world);
+        else log_io("e~Failed to call function '%s' in library", ss.str().c_str());
     }
 }
 
@@ -200,7 +211,12 @@ World* LoadWorldFromData(Application& app, const WorldFileData& data)
         return nullptr;
     }
 
-    PassContextsToDll(lib, app);
+    bool success = PassContextsToDll(lib, app);
+
+    if (!success) // exit on no context pass
+    {
+        return nullptr;
+    }
 
     // Create the world, we need this for the factory functions
     //                   because the world needs their type info
